// src/app/shared/models/image.ts
/**
* Class with Image data needed for the image gallery.
*
* The src property stores the image source, the name stores the name of the image
* and description adds an additional description.
*
* @author Jose Danado (jose.danado@gmail.com)
*/
export class Image {
  name: string;
  description: string;
  src: string;
}